Revert "W06 rama dashboard backend"
