+++ auth.config.ts
@@ -7,10 +7,9 @@ export const authConfig = {
   callbacks: {
     authorized({ auth, request: { nextUrl } }) {
       const isLoggedIn = !!auth?.user;
-
-      const isProtectedRoute = nextUrl.pathname.startsWith('/dashboard') || nextUrl.pathname.startsWith('/product') || nextUrl.pathname.startsWith('/profile');
-
-      if (isProtectedRoute) {
+      const isOnDashboard = nextUrl.pathname.startsWith('/dashboard');
+      const isOnProductPage = nextUrl.pathname.startsWith('/product');
+      if (isOnDashboard || isOnProductPage) {
         if (isLoggedIn) return true;
         return false; // Redirect unauthenticated users to login page
       } else if (isLoggedIn) {
