+++ package.json
@@ -21,13 +21,11 @@
     "next-connect": "^0.12.1",
     "react": "^18",
     "react-dom": "^18",
-    "use-debounce": "^10.0.1",
     "uuid": "^10.0.0",
     "zod": "^3.23.8"
   },
   "devDependencies": {
     "@types/bcrypt": "^5.0.2",
-    "@types/next-auth": "^3.15.0",
     "@types/multer": "^1.4.11",
     "@types/node": "^20",
     "@types/react": "^18",
