+++ src/app/lib/data.ts
@@ -1,5 +1,5 @@
-import { User, Product, Category } from './definitions';
 import { revalidatePath } from 'next/cache';
+import { User, Products } from './definitions';
 import { sql } from '@vercel/postgres';
 
 
@@ -41,107 +41,6 @@ export async function fetchFilteredUserDetails(
     }
   }
 
-export async function UpdateUserDetails(
-  userId: string,
-) {
-
-  try {
-    await sql`
-    UPDATE users SET profile_image='carol_profile.jpg' 
-    WHERE id=${userId}`;
-
-  } catch (error) {
-    console.error('Error updating or fetching user details:', error);
-      throw error; // Propagate the error to the caller
-  }
-}
-
-export async function fetchFilteredProductsByUser(
-  userId: string,
-) {
-  try {
-    const data = await sql<Product>`
-      SELECT products.id, products.user_id, products.product_name, products.category, products.price, products.image_url, products.description, products.stock
-      FROM users
-      JOIN products ON users.id = products.user_id
-      WHERE 
-      users.id=${userId}`;
-
-    //const latestInvoices = data.rows.map((invoice) => ({
-      //...invoice,
-      // amount: formatCurrency(invoice.amount),
-    // }));
-    return data.rows;
-  } catch (error) {
-    console.error('Database Error:', error);
-    throw new Error('Failed to fetch the products by user.');
-  }
-}
-
-export async function getProducts() {
-  try {
-    const data = await sql<Product>`
-      SELECT *
-      FROM products
-    `;
-    const products = data.rows;
-    return products;
-  } catch (err) {
-    console.error('Database Error:', err);
-    throw new Error('Failed to fetch all products.');
-  }
-}
-
-export async function getCategories() {
-  try {
-    const data = await sql<Category>`
-      SELECT DISTINCT category
-      FROM products
-    `;
-    const categories = data.rows;
-    return categories;
-  } catch (err) {
-    console.error('Database Error:', err);
-    throw new Error('Failed to fetch all categories.');
-  }
-}
-
-export async function getFilteredProducts(
-  query: string,
-) {
-  try {
-    const products = await sql<Product>`
-      SELECT *
-      FROM products
-      WHERE
-        product_name ILIKE ${`%${query}%`} OR
-        description ILIKE ${`%${query}%`} OR
-        category ILIKE ${`%${query}%`}
-    `;
-
-    return products.rows;
-  } catch (error) {
-    console.error('Database Error:', error);
-    throw new Error('Failed to fetch products.');
-  }
-}
-
-export async function getProductById(id: string) {
-  try {
-    const products = await sql<Product>`
-      SELECT *
-      FROM products
-      WHERE
-        id = ${id}
-    `;
-
-    return products.rows[0];
-  } catch (error) {
-    console.error('Database Error:', error);
-    throw new Error('Failed to fetch products.');
-  }
-}
-=======
   export async function fetchFilteredProductsByUser(
     userId: string,
   ) {
